+++ app_sources/Inputs/Gamepad/directive.js
@@ -17,7 +17,6 @@ app.directive('gamepad', ['gamepads', function (gamepads) {
             action    : attrs.keyboardaction,
           });
       });
-
     }
   };
 }]);